requirements.txt
# pip install -r requirements.txt
python-dotenv
anthropic
anthropic[bedrock]
anthropic[vertex]
google-cloud-aiplatform
nbformat>=5.7.0
boto3>=1.35.23
botocore>=1.35.23
tabulate
pylance
types-boto3

# Optional but recommended for better IDE support
boto3-stubs
mypy-boto3-bedrock
mypy-boto3-iam
mypy-boto3-sts
